+++ scripts/doctest.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+set -euo pipefail
+
+cabal repl --build-depends=unliftio --with-compiler=doctest --repl-options='-w -Wdefault'